pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swarm-simulation"
version = "0.1.0"
description = "A generalized swarm intelligence simulation engine using knowledge graphs and multi-agent systems"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "networkx>=3.1",
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "camel-ai>=0.2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "ollama>=0.3.0",
]

[project.optional-dependencies]
neo4j = ["neo4j>=5.0"]
gemini = ["google-generativeai>=0.5.0"]
ui = ["fastapi>=0.110", "uvicorn>=0.29", "websockets>=12.0"]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.10"]
all = ["swarm-simulation[neo4j,gemini,ui,dev]"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
